Eliminato file di test, inserito in sniffer.py il codice testato su Raspberry
